Add required imports and remove duplicate cast

+++ FundamentalsCore/src/com/johnymuffin/beta/fundamentals/commands/CommandBalance.java
@@ -62,7 +62,7 @@ public class CommandBalance implements CommandExecutor {
     private void printBalance(String playerName, CommandSender commandSender, EconomyAPI.BalanceWrapper balanceWrapper) {
         if (balanceWrapper.getEconomyResult() == EconomyAPI.EconomyResult.successful) {
             String message = FundamentalsLanguage.getInstance().getMessage("balance_successful");
-            message = message.replace("%var1%", String.valueOf((double) Utils.round(balanceWrapper.getBalance(), 2)));
+            message = message.replace("%var1%", String.valueOf(Utils.round(balanceWrapper.getBalance(), 2)));
             commandSender.sendMessage(message);
         } else if (balanceWrapper.getEconomyResult() == EconomyAPI.EconomyResult.userNotKnown) {
             String message = FundamentalsLanguage.getInstance().getMessage("player_not_found_full");
